refactor: rework proposal exports and api
