+++ app/features/event-management/proposals-export/json.tsx
@@ -6,15 +6,16 @@ import {
   requireAuthorizedTeam,
 } from '~/shared/authorization/authorization.middleware.ts';
 import type { Route } from './+types/json.ts';
-import { CfpReviewsExports } from './services/cfp-reviews-exports.server.ts';
+import { ProposalsExport } from './services/proposals-export.server.ts';
 
 export const middleware = [requireAuth, requireAuthorizedTeam, requireAuthorizedEvent];
 
 export const loader = async ({ request, params, context }: Route.LoaderArgs) => {
   const authorizedEvent = context.get(AuthorizedEventContext);
   const filters = parseUrlFilters(request.url);
-  const exports = CfpReviewsExports.for(authorizedEvent);
-  const results = await exports.forJson(filters);
+
+  const results = await ProposalsExport.forUser(authorizedEvent).toJson(filters);
+
   return Response.json(results, {
     headers: { 'Content-Disposition': `attachment; filename="${params.event}-proposals.json"` },
   });
